pyproject.toml
[project]
name = "alpha"
version = "0.1.0"
description = "Event-driven + deep-value research system for asymmetric equity bets"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-dateutil>=2.9",
    "lxml>=5.2",
    "beautifulsoup4>=4.12",
    "anthropic>=0.40",
    "rich>=13.7",
    "typer>=0.12",
    "tenacity>=8.3",
    "pandas>=2.2",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
alpha = "alpha.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["alpha*"]

[tool.ruff]
line-length = 100
target-version = "py311"
